backend/requirements.txt
django
djangorestframework
python-dotenv
pyjwt